Move GlobalBlocksTypedFormMetadataVisitor to AdminBundle
